snowflake_requirements.txt
# Snowflake MCP Server Dependencies

# Core MCP framework
mcp>=1.0.0

# Snowflake connector
snowflake-connector-python>=3.6.0

# Cryptography for key pair authentication
cryptography>=41.0.0

# Configuration parsing
pyyaml>=6.0.1

# Type hints support
typing-extensions>=4.8.0

# Optional dependencies for enhanced functionality
# Uncomment as needed:

# For advanced SQL operations
# snowflake-sqlalchemy>=1.5.0

# For data frame operations
# pandas>=2.0.0

# For optimized data transfer
# pyarrow>=14.0.0

# For metrics collection
# prometheus-client>=0.19.0

# Development dependencies (for testing and development)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-mock>=3.12.0
# black>=23.0.0
# mypy>=1.7.0